% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_bf_onestage_design.R
\name{as.data.frame.singlearm_onestage_bf_design}
\alias{as.data.frame.singlearm_onestage_bf_design}
\title{Convert a one-stage single-arm BF design to a data frame}
\usage{
\method{as.data.frame}{singlearm_onestage_bf_design}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"singlearm_onestage_bf_design"}.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Currently unused.}
}
\value{
A data frame with the search results.
}
\description{
Convert a one-stage single-arm BF design to a data frame
}
