% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_rope_twostage.R
\name{as.data.frame.singlearm_rope_twostage_design}
\alias{as.data.frame.singlearm_rope_twostage_design}
\title{Convert candidates to a data frame}
\usage{
\method{as.data.frame}{singlearm_rope_twostage_design}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"singlearm_rope_twostage_design"}.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Further arguments (currently ignored).}
}
\value{
The candidates data frame.
}
\description{
Convert candidates to a data frame
}
