% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfbin2arm-package.R
\docType{package}
\name{bfbin2arm}
\alias{bfbin2arm-package}
\alias{bfbin2arm}
\title{bfbin2arm: Bayesian Power for Two-Arm Binomial Bayes Factors}
\description{
Functions for Bayesian power and sample size calculation in two-arm binomial
trials using Bayes factors for point-null and directional hypotheses.
}
\details{
The package provides simulation-free and simulation-based methods to
calibrate Bayes factor designs in two-arm phase II trials with binary
endpoints, including power and type-I-error calculations and visualization
tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rikokelter.github.io/bfbin2arm/}
  \item \url{https://github.com/rikokelter/bfbin2arm}
  \item \url{https://arxiv.org/abs/2511.23144}
  \item \url{https://arxiv.org/abs/2603.01715}
  \item \url{https://arxiv.org/abs/2502.02914}
  \item \url{https://arxiv.org/abs/2606.02410}
  \item \url{https://rikokelter.github.io/bfbin2arm/}
}

}
\author{
\strong{Maintainer}: Riko Kelter \email{rkelter@uni-koeln.de}

}
\keyword{internal}
