% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{BFminus0}
\alias{BFminus0}
\title{Bayes factor BF-0: H- vs H0}
\usage{
BFminus0(BFminus1, BF01)
}
\arguments{
\item{BFminus1}{Value of BF\eqn{_{-1}}.}

\item{BF01}{Value of BF\eqn{_{01}}.}
}
\value{
Numeric scalar, BF\eqn{_{-0}}.
}
\description{
Bayes factor BF-0: H- vs H0
}
