% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{BFminus1}
\alias{BFminus1}
\title{Bayes factor BF-1: H- vs H1}
\usage{
BFminus1(y1, y2, n1, n2, a_1_d = 1, b_1_d = 1, a_2_d = 1, b_2_d = 1)
}
\arguments{
\item{y1}{Number of successes in arm 1 (control).}

\item{y2}{Number of successes in arm 2 (treatment).}

\item{n1}{Sample size in arm 1.}

\item{n2}{Sample size in arm 2.}

\item{a_1_d, b_1_d}{Design-prior shape parameters for \eqn{p_1}.}

\item{a_2_d, b_2_d}{Design-prior shape parameters for \eqn{p_2}.}
}
\value{
Numeric scalar, BF\eqn{_{-1}}.
}
\description{
Bayes factor BF-1: H- vs H1
}
