% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{BFplus1}
\alias{BFplus1}
\title{Bayes factor BF+1: H+ vs H1}
\usage{
BFplus1(y1, y2, n1, n2, a_1_a = 1, b_1_a = 1, a_2_a = 1, b_2_a = 1)
}
\arguments{
\item{y1}{Number of successes in arm 1 (control).}

\item{y2}{Number of successes in arm 2 (treatment).}

\item{n1}{Sample size in arm 1.}

\item{n2}{Sample size in arm 2.}

\item{a_1_a, b_1_a}{Shape parameters of the Beta prior for \eqn{p_1} under the
alternative (analysis prior).}

\item{a_2_a, b_2_a}{Shape parameters of the Beta prior for \eqn{p_2} under the
alternative (analysis prior).}
}
\value{
Numeric scalar, BF\eqn{_{+1}} = posterior odds / prior odds for H+ vs H1.
}
\description{
Bayes factor BF+1: H+ vs H1
}
