% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{BFplusMinus}
\alias{BFplusMinus}
\title{Bayes factor BF+-: H+ vs H-}
\usage{
BFplusMinus(BFplus1, BFminus1)
}
\arguments{
\item{BFplus1}{Value of BF\eqn{_{+1}}.}

\item{BFminus1}{Value of BF\eqn{_{-1}}.}
}
\value{
Numeric scalar, BF\eqn{_{+-}}.
}
\description{
Bayes factor BF+-: H+ vs H-
}
