% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_rope_twostage.R
\name{coef.singlearm_rope_twostage_design}
\alias{coef.singlearm_rope_twostage_design}
\title{Extract design coefficients}
\usage{
\method{coef}{singlearm_rope_twostage_design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"singlearm_rope_twostage_design"}.}

\item{...}{Further arguments (currently ignored).}
}
\value{
A named numeric vector of key design quantities.
}
\description{
Extract design coefficients
}
