% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_oc_twoarm_helpers.R
\name{freq_oc_twostage_twoarm_fixed}
\alias{freq_oc_twostage_twoarm_fixed}
\title{Exact frequentist operating characteristics for a fixed two-stage two-arm BF design}
\usage{
freq_oc_twostage_twoarm_fixed(
  n1_1,
  n1_2,
  n2_1,
  n2_2,
  k,
  k_f,
  test,
  p1,
  p2,
  a_0_a,
  b_0_a,
  a_1_a,
  b_1_a,
  a_2_a,
  b_2_a,
  a_1_a_Hminus,
  b_1_a_Hminus,
  a_2_a_Hminus,
  b_2_a_Hminus
)
}
\description{
Internal helper. Computes the exact frequentist rejection probability of the
two-stage design with futility stopping at interim and efficacy decision at final.
}
\keyword{internal}
