% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-twoarm_rope_onestage.R
\name{plot.bfbin2armrope2armdesign}
\alias{plot.bfbin2armrope2armdesign}
\title{Plot a two-arm one-stage ROPE equivalence design}
\usage{
\method{plot}{bfbin2armrope2armdesign}(
  x,
  type = c("overview", "operatingcharacteristics", "heatmap", "prior"),
  ...
)
}
\arguments{
\item{x}{A \code{bfbin2armrope2armdesign} object.}

\item{type}{Plot type.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot a two-arm one-stage ROPE equivalence design
}
