% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{postProbHminus}
\alias{postProbHminus}
\title{Posterior probability P(p2 <= p1 | data)}
\usage{
postProbHminus(y1, y2, n1, n2, a_1_a = 1, b_1_a = 1, a_2_a = 1, b_2_a = 1)
}
\arguments{
\item{y1}{Number of successes in arm 1 (control).}

\item{y2}{Number of successes in arm 2 (treatment).}

\item{n1}{Sample size in arm 1.}

\item{n2}{Sample size in arm 2.}

\item{a_1_a, b_1_a}{Shape parameters of the Beta prior for \eqn{p_1} under the
alternative (analysis prior).}

\item{a_2_a, b_2_a}{Shape parameters of the Beta prior for \eqn{p_2} under the
alternative (analysis prior).}
}
\value{
Numeric scalar, posterior probability \eqn{P(p_2 <= p_1 | y_1, y_2)}.
}
\description{
Posterior probability P(p2 <= p1 | data)
}
