% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{postProbHplus}
\alias{postProbHplus}
\title{Posterior probability P(p2 > p1 | data) under independent Beta priors}
\usage{
postProbHplus(y1, y2, n1, n2, a_1_d = 1, b_1_d = 1, a_2_d = 1, b_2_d = 1)
}
\arguments{
\item{y1}{Number of successes in arm 1 (control).}

\item{y2}{Number of successes in arm 2 (treatment).}

\item{n1}{Sample size in arm 1.}

\item{n2}{Sample size in arm 2.}

\item{a_1_d, b_1_d}{Design-prior shape parameters for \eqn{p_1}.}

\item{a_2_d, b_2_d}{Design-prior shape parameters for \eqn{p_2}.}
}
\value{
Numeric scalar, posterior probability \eqn{P(p_2 > p_1 | y_1, y_2)}.
}
\description{
Posterior probability P(p2 > p1 | data) under independent Beta priors
}
