% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_densities.R
\name{predictiveDensityH0}
\alias{predictiveDensityH0}
\title{Predictive density under H0: p1 = p2 = p}
\usage{
predictiveDensityH0(y1, y2, n1, n2, a_0_d = 1, b_0_d = 1)
}
\arguments{
\item{y1}{Number of successes in arm 1 (control).}

\item{y2}{Number of successes in arm 2 (treatment).}

\item{n1}{Sample size in arm 1.}

\item{n2}{Sample size in arm 2.}

\item{a_0_d, b_0_d}{Design-prior parameters for common p under H0.}
}
\value{
Numeric scalar, predictive density.
}
\description{
Beta-binomial predictive density for data (y1,y2) under H0.
}
