% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_densities.R
\name{predictiveDensityH1}
\alias{predictiveDensityH1}
\title{Predictive density under H1: p1 != p2}
\usage{
predictiveDensityH1(y1, y2, n1, n2, a_1_d = 1, b_1_d = 1, a_2_d = 1, b_2_d = 1)
}
\arguments{
\item{y1}{Number of successes in arm 1 (control).}

\item{y2}{Number of successes in arm 2 (treatment).}

\item{n1}{Sample size in arm 1.}

\item{n2}{Sample size in arm 2.}

\item{a_1_d, b_1_d}{Design-prior parameters for p1.}

\item{a_2_d, b_2_d}{Design-prior parameters for p2.}
}
\value{
Numeric scalar, predictive density.
}
\description{
Product of two independent Beta-binomial predictive densities.
}
