% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_bf_onestage_design.R
\name{print.singlearm_onestage_bf_design}
\alias{print.singlearm_onestage_bf_design}
\title{Print method for one-stage single-arm BF designs}
\usage{
\method{print}{singlearm_onestage_bf_design}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"singlearm_onestage_bf_design"}.}

\item{...}{Currently unused.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for one-stage single-arm BF designs
}
