% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_bf_design.R
\name{print.summary.singlearm_bf_design}
\alias{print.summary.singlearm_bf_design}
\title{Print method for summary.singlearm_bf_design}
\usage{
\method{print}{summary.singlearm_bf_design}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.singlearm_bf_design".}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for summary.singlearm_bf_design
}
