% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_bf_onestage_design.R
\name{print.summary.singlearm_onestage_bf_design}
\alias{print.summary.singlearm_onestage_bf_design}
\title{Print method for summaries of one-stage single-arm BF designs}
\usage{
\method{print}{summary.singlearm_onestage_bf_design}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.singlearm_onestage_bf_design"}.}

\item{digits}{Number of digits to print.}

\item{...}{Currently unused.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print method for summaries of one-stage single-arm BF designs
}
