% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_rope_twostage.R
\name{print.summary.singlearm_rope_twostage_design}
\alias{print.summary.singlearm_rope_twostage_design}
\title{Print a compact summary of a single-arm two-stage ROPE design}
\usage{
\method{print}{summary.singlearm_rope_twostage_design}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.singlearm_rope_twostage_design"}.}

\item{digits}{Number of digits used to display operating characteristics.}

\item{...}{Further arguments (currently ignored).}
}
\description{
Print a compact summary of a single-arm two-stage ROPE design
}
