% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{priorProbHminus}
\alias{priorProbHminus}
\title{Prior probability P(p2 <= p1) under independent Beta priors}
\usage{
priorProbHminus(a_1_d, b_1_d, a_2_d, b_2_d)
}
\arguments{
\item{a_1_d, b_1_d}{Shape parameters of the Beta prior for \eqn{p_1}.}

\item{a_2_d, b_2_d}{Shape parameters of the Beta prior for \eqn{p_2}.}
}
\value{
Numeric scalar, prior probability P(p2 <= p1).
}
\description{
Prior probability P(p2 <= p1) under independent Beta priors
}
