% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{priorProbHplus}
\alias{priorProbHplus}
\title{Prior probability P(p2 > p1) under independent Beta priors}
\usage{
priorProbHplus(a_1_a, b_1_a, a_2_a, b_2_a)
}
\arguments{
\item{a_1_a, b_1_a}{Shape parameters of the Beta prior for \eqn{p_1}.}

\item{a_2_a, b_2_a}{Shape parameters of the Beta prior for \eqn{p_2}.}
}
\value{
Numeric scalar, prior probability P(p2 > p1).
}
\description{
Prior probability P(p2 > p1) under independent Beta priors
}
