% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_bf_design.R
\name{summary.singlearm_bf_design}
\alias{summary.singlearm_bf_design}
\title{Summary for single-arm BF designs}
\usage{
\method{summary}{singlearm_bf_design}(object, ...)
}
\arguments{
\item{object}{An object of class "singlearm_bf_design".}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class "summary.singlearm_bf_design".
}
\description{
Summary for single-arm BF designs
}
