% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_bf_onestage_design.R
\name{summary.singlearm_onestage_bf_design}
\alias{summary.singlearm_onestage_bf_design}
\title{Summarize a one-stage single-arm BF design}
\usage{
\method{summary}{singlearm_onestage_bf_design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"singlearm_onestage_bf_design"}.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{"summary.singlearm_onestage_bf_design"}.
}
\description{
Summarize a one-stage single-arm BF design
}
