% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_rope_twostage.R
\name{summary.singlearm_rope_twostage_design}
\alias{summary.singlearm_rope_twostage_design}
\title{Summary for a single-arm two-stage ROPE design}
\usage{
\method{summary}{singlearm_rope_twostage_design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"singlearm_rope_twostage_design"}.}

\item{...}{Further arguments (currently ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary for a single-arm two-stage ROPE design
}
