% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-singlearm_rope_twostage.R
\name{print.singlearm_rope_twostage_design}
\alias{print.singlearm_rope_twostage_design}
\title{Print a single-arm two-stage ROPE design}
\usage{
\method{print}{singlearm_rope_twostage_design}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"singlearm_rope_twostage_design"}.}

\item{...}{Further arguments (currently ignored).}

\item{digits}{Number of digits to display. Default \code{4L}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a single-arm two-stage ROPE design
}
